+++ src/usettings.pas
@@ -15,12 +15,15 @@ type
   TFSettings = class(TForm)
     BPDefault: TButtonPanel;
     LEHexEditor: TLabeledEdit;
+    LECompileCommand: TLabeledEdit;
     OPSelectFile: TOpenDialog;
     SPChooseDirectory: TSpeedButton;
+    SPChooseDirectory1: TSpeedButton;
     procedure CancelButtonClick(Sender: TObject);
+    procedure ChooseCompileCommand(Sender: TObject);
     procedure FormShow(Sender: TObject);
     procedure OKButtonClick(Sender: TObject);
-    procedure SPChooseDirectoryClick(Sender: TObject);
+    procedure ChooseHexEditor(Sender: TObject);
   private
 
   public
@@ -41,18 +44,26 @@ begin
   Close;
 end;
 
+procedure TFSettings.ChooseCompileCommand(Sender: TObject);
+begin
+  if OPSelectFile.Execute then
+    LECompileCommand.Caption := OPSelectFile.FileName;
+end;
+
 procedure TFSettings.FormShow(Sender: TObject);
 begin
   LEHexEditor.Caption := ConfigObj.HexEditor;
+  LECompileCommand.Caption := ConfigObj.CompileCommand;
 end;
 
 procedure TFSettings.OKButtonClick(Sender: TObject);
 begin
   ConfigObj.HexEditor := LEHexEditor.Caption;
+  ConfigObj.CompileCommand := LECompileCommand.Caption;
   Close;
 end;
 
-procedure TFSettings.SPChooseDirectoryClick(Sender: TObject);
+procedure TFSettings.ChooseHexEditor(Sender: TObject);
 begin
   if OPSelectFile.Execute then
     LEHexEditor.Caption := OPSelectFile.FileName;